requirements.txt
python-dotenv
icecream

requests
streamlit
watchdog
pycountry

surrealdb
